WIP change of name and api
