+++ apps/docs/package.json
@@ -14,14 +14,14 @@
     "with-env": "dotenv -e ../../.env --"
   },
   "dependencies": {
-    "@astrojs/check": "^0.3.1",
-    "@astrojs/starlight": "^0.12.0",
+    "@astrojs/check": "^0.5.5",
+    "@astrojs/starlight": "^0.19.1",
     "@astrojs/starlight-tailwind": "^2.0.1",
-    "@astrojs/tailwind": "^5.0.0",
-    "astro": "^3.4.3",
+    "@astrojs/tailwind": "^5.1.0",
+    "astro": "^4.4.3",
     "dotenv-cli": "^7.3.0",
-    "sharp": "^0.32.5",
-    "tailwindcss": "^3",
-    "typescript": "^5.2.2"
+    "sharp": "^0.33.2",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3"
   }
 }
