+++ apps/docs/src/content/docs/getting-started/nextjs.md
@@ -1,0 +1,10 @@
+---
+title: Installing for Next.js applications.
+description: How to install the `@sp-hooks` packages in a Next.js application.
+---
+
+## Installation
+
+```sh
+npm install @sp-hooks/next
+```
