+++ apps/docs/src/content/docs/getting-started/react.md
@@ -1,0 +1,10 @@
+---
+title: Installing for generic React applications.
+description: How to install the `@sp-hooks` packages in a generic React application.
+---
+
+## Installation
+
+```sh
+npm install @sp-hooks/react
+```
