+++ apps/docs/src/content/docs/guides/rsc.md
@@ -1,0 +1,4 @@
+---
+title: Usage with React Server Components
+description: Everything you need to know about using sp-hooks with React Server Components.
+---
