+++ apps/docs/src/content/docs/guides/validation.md
@@ -1,0 +1,4 @@
+---
+title: State validation when parsing from URLSearchParams
+description: How to validate state when parsing from URLSearchParams
+---
