+++ apps/docs/src/content/docs/index.mdx
@@ -1,6 +1,6 @@
 ---
-title: Welcome to Starlight
-description: Get started building your docs site with Starlight.
+title: Search Params Hooks
+description: Hooks to sync and manage state with URL search params.
 template: splash
 hero:
   title: |
