+++ apps/docs/tailwind.config.mjs
@@ -8,7 +8,7 @@ export default {
     extend: {
       colors: {
         // Your preferred accent color. Indigo is closest to Starlight’s defaults.
-        accent: colors.indigo,
+        accent: colors.orange,
         // Your preferred gray scale. Zinc is closest to Starlight’s defaults.
         gray: colors.zinc,
       },
