+++ apps/playground/next.config.mjs
@@ -1,7 +1,7 @@
 export default {
   reactStrictMode: true,
   transpilePackages: [
-    "@use-search-params-state/react",
-    "@use-search-params-state/next",
+    "@sp-hooks/react",
+    "@sp-hooks/next",
   ],
 };
