+++ apps/playground/package.json
@@ -13,14 +13,14 @@
   },
   "eslintConfig": {
     "extends": [
-      "@use-search-params-state/eslint-config/base",
-      "@use-search-params-state/eslint-config/nextjs",
-      "@use-search-params-state/eslint-config/react"
+      "@sp-hooks/eslint-config/base",
+      "@sp-hooks/eslint-config/nextjs",
+      "@sp-hooks/eslint-config/react"
     ],
     "root": true
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.3.2",
+    "@hookform/resolvers": "^3.3.4",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-aspect-ratio": "^1.0.3",
@@ -48,37 +48,37 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@t3-oss/env-nextjs": "^0.7.1",
-    "@use-search-params-state/next": "workspace:*",
-    "@use-search-params-state/react": "workspace:*",
-    "@vercel/analytics": "^1.1.1",
+    "@t3-oss/env-nextjs": "^0.7.3",
+    "@sp-hooks/next": "workspace:*",
+    "@sp-hooks/react": "workspace:*",
+    "@vercel/analytics": "^1.2.2",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
-    "cmdk": "^0.2.0",
-    "date-fns": "^2.30.0",
+    "clsx": "^2.1.0",
+    "cmdk": "^0.2.1",
+    "date-fns": "^3.3.1",
     "dotenv-cli": "^7.3.0",
-    "eslint": "^8",
-    "lucide-react": "^0.292.0",
+    "eslint": "^8.56.0",
+    "lucide-react": "^0.338.0",
     "next-themes": "^0.2.1",
-    "react-day-picker": "^8.9.1",
+    "react-day-picker": "^8.10.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.47.0",
-    "tailwind-merge": "^2.0.0",
+    "react-hook-form": "^7.50.1",
+    "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/node": "^18",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "@use-search-params-state/eslint-config": "workspace:*",
-    "@use-search-params-state/tsconfig": "workspace:*",
-    "autoprefixer": "^10.4.16",
+    "@types/node": "^18.19.18",
+    "@types/react": "^18.2.58",
+    "@types/react-dom": "^18.2.19",
+    "@sp-hooks/eslint-config": "workspace:*",
+    "@sp-hooks/tsconfig": "workspace:*",
+    "autoprefixer": "^10.4.17",
     "docs": "^0.0.1",
-    "next": "^14.0.1",
-    "postcss": "^8.4.31",
+    "next": "^14.1.0",
+    "postcss": "^8.4.35",
     "react": "^18.2.0",
-    "tailwindcss": "^3",
-    "typescript": "^5.2.2"
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3"
   }
 }
