+++ apps/playground/src/app/array-values/page.tsx
@@ -1,12 +1,14 @@
 "use client";
 
+import { useState } from "react";
 import { SetKeyValueArrayInputs } from "@/components/set-key-value-inputs";
 import { Alert } from "@/components/ui/alert";
-
-import { useSearchParamsState } from "@use-search-params-state/next";
+import { useObserverAndSearch } from "@sp-hooks/next";
 
 export default function Page() {
-  const [state, setState] = useSearchParamsState();
+  const [state, setState] = useState();
+
+  useObserverAndSearch(state);
 
   return (
     <div>
@@ -14,7 +16,7 @@ export default function Page() {
         {"This demonstration shows how to use the library with arrays."}
       </Alert>
 
-      <SetKeyValueArrayInputs setState={(k, v) => setState(k, v)} />
+      <SetKeyValueArrayInputs setState={setState} />
 
       <pre>{JSON.stringify(state, null, 2)}</pre>
     </div>
