+++ packages/next/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @use-search-params-state/next
+# @sp-hooks/next
 
 ## 0.0.5
 
@@ -6,7 +6,7 @@
 
 - 2a8964c: Added array values feature
 - Updated dependencies [2a8964c]
-  - @use-search-params-state/react@0.0.5
+  - @sp-hooks/react@0.0.5
 
 ## 0.0.4
 
@@ -14,7 +14,7 @@
 
 - c6f9574: Added generic type parameter to the useSearchParamsState hooks
 - Updated dependencies [c6f9574]
-  - @use-search-params-state/react@0.0.4
+  - @sp-hooks/react@0.0.4
 
 ## 0.0.3
 
@@ -22,7 +22,7 @@
 
 - 51764f2: Fixed peer dependencies versions
 - Updated dependencies [51764f2]
-  - @use-search-params-state/react@0.0.3
+  - @sp-hooks/react@0.0.3
 
 ## 0.0.2
 
@@ -30,7 +30,7 @@
 
 - aca085d: Adde sortKeys option, with default set to true.
 - Updated dependencies [aca085d]
-  - @use-search-params-state/react@0.0.2
+  - @sp-hooks/react@0.0.2
 
 ## 0.0.1
 
@@ -38,4 +38,4 @@
 
 - c81587f: use-search-params-state mvp release
 - Updated dependencies [c81587f]
-  - @use-search-params-state/react@0.0.1
+  - @sp-hooks/react@0.0.1
