+++ packages/next/src/index.ts
@@ -1,24 +1,36 @@
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import {
+  useObserveAndStore as useObserveAndStoreReact,
+  type SupportedValues,
+} from "@sp-hooks/react";
 
-import { useSearchParamsState as useSearchParamsStateReact } from "@use-search-params-state/react";
-import type { UseSearchParamsStateOptions } from "@use-search-params-state/react";
+// export const useSearchParamsState = <
+//   S extends Partial<Record<string, string | string[]>>,
+// >(
+//   opts?: UseSearchParamsStateOptions<S>,
+// ) => {
+//   const searchParams = useSearchParams();
+//   const router = useRouter();
+//   const pathname = usePathname();
 
-export const useSearchParamsState = <
-  S extends Partial<Record<string, string | string[]>>,
->(
-  opts?: UseSearchParamsStateOptions<S>,
-) => {
-  const searchParams = useSearchParams();
+//   const setSearchParams = (newSearchParams: URLSearchParams) => {
+//     router.push(pathname + "?" + newSearchParams.toString());
+//   };
+
+//   return useSearchParamsStateReact<S>({
+//     searchParams,
+//     setSearchParams,
+//     ...opts,
+//   });
+// };
+
+export function useObserveAndStore<S extends Record<string, SupportedValues>>(
+  state: S,
+) {
   const router = useRouter();
   const pathname = usePathname();
 
-  const setSearchParams = (newSearchParams: URLSearchParams) => {
+  useObserveAndStoreReact(state, (newSearchParams) => {
     router.push(pathname + "?" + newSearchParams.toString());
-  };
-
-  return useSearchParamsStateReact<S>({
-    searchParams,
-    setSearchParams,
-    ...opts,
   });
-};
+}
