+++ tooling/github/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "@use-search-params-state/github",
+  "name": "@sp-hooks/github",
   "private": true
 }
