+++ tooling/prettier/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@use-search-params-state/prettier-config",
+  "name": "@sp-hooks/prettier-config",
   "private": true,
   "version": "0.1.1",
   "main": "index.mjs",
@@ -9,11 +9,11 @@
   },
   "dependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
-    "prettier": "^3.0.3",
-    "prettier-plugin-tailwindcss": "^0.5.6"
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11"
   },
   "devDependencies": {
-    "@use-search-params-state/tsconfig": "workspace:*",
-    "typescript": "^5.2.2"
+    "@sp-hooks/tsconfig": "workspace:*",
+    "typescript": "^5.3.3"
   }
 }
